backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Romain Mularczyk <romain.mularczyk@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.87.0"
uvicorn = "^0.20.0"
pymongo = "^4.3.3"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.10.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
